Extract `build_property_field` into a base class
